backend/pyproject.toml
[project]
name = "kino-mcp-project"
version = "0.1.0"
description = "sample demo of MCP server with OAuth"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.10.6",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.41",
    "starlette>=0.37.0",
    "stytch>=13.12.0",
]
